requirements.txt
openai
langchain
langchain-openai
chainlit
python-dotenv
